+++ api/src/main/java/io/strimzi/api/kafka/model/storage/EphemeralStorage.java
@@ -2,44 +2,46 @@
  * Copyright 2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.api.kafka.model;
+package io.strimzi.api.kafka.model.storage;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import io.strimzi.crdgenerator.annotations.Description;
 import io.strimzi.crdgenerator.annotations.Minimum;
+import io.sundr.builder.annotations.Buildable;
 import lombok.EqualsAndHashCode;
 
-@JsonTypeInfo(
-        use = JsonTypeInfo.Id.NAME,
-        include = JsonTypeInfo.As.EXISTING_PROPERTY,
-        property = "type"
-)
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = EphemeralStorage.class, name = Storage.TYPE_EPHEMERAL),
-        @JsonSubTypes.Type(value = PersistentClaimStorage.class, name = Storage.TYPE_PERSISTENT_CLAIM)}
+/**
+ * Representation for ephemeral storage.
+ */
+@Buildable(
+        editableEnabled = false,
+        generateBuilderPackage = false,
+        builderPackage = "io.fabric8.kubernetes.api.builder"
 )
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @EqualsAndHashCode
-public abstract class SingleVolumeStorage extends Storage {
+public class EphemeralStorage extends SingleVolumeStorage {
 
     private static final long serialVersionUID = 1L;
 
     private Integer id;
 
+    @Description("Must be `" + TYPE_EPHEMERAL + "`")
     @Override
-    @Description("Storage type, must be either 'ephemeral' or 'persistent-claim'.")
-    public abstract String getType();
+    public String getType() {
+        return TYPE_EPHEMERAL;
+    }
 
+    @Override
     @Description("Storage identification number. It is mandatory only for storage volumes defined in a storage of type 'jbod'")
     @Minimum(0)
     @JsonInclude(JsonInclude.Include.NON_NULL)
     public Integer getId() {
-        return id;
+        return super.getId();
     }
 
+    @Override
     public void setId(Integer id) {
-        this.id = id;
+        super.setId(id);
     }
 }