+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperator.java
@@ -29,7 +29,7 @@ import io.strimzi.api.kafka.model.CertificateAuthority;
 import io.strimzi.api.kafka.model.DoneableKafka;
 import io.strimzi.api.kafka.model.ExternalLogging;
 import io.strimzi.api.kafka.model.Kafka;
-import io.strimzi.api.kafka.model.Storage;
+import io.strimzi.api.kafka.model.storage.Storage;
 import io.strimzi.certs.CertManager;
 import io.strimzi.operator.cluster.ClusterOperator;
 import io.strimzi.operator.cluster.ClusterOperatorConfig;
@@ -149,7 +149,6 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
         this.podOperations = supplier.podOperations;
         this.ingressOperations = supplier.ingressOperations;
         this.storageClassOperator = supplier.storageClassOperations;
-
     }
 
     @Override
@@ -1584,7 +1583,7 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                             result.complete();
                         } else {
                             // The PVC is Bound and resizing is not in progress => We should check if the SC supports resizing and check if size changed
-                            Quantity currentSize = currentPvc.getStatus().getCapacity().get("storage");
+                            Quantity currentSize = currentPvc.getSpec().getResources().getRequests().get("storage");
                             Quantity desiredSize = desiredPvc.getSpec().getResources().getRequests().get("storage");
 
                             if (!desiredSize.equals(currentSize))   {
@@ -1719,7 +1718,7 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                     List<PersistentVolumeClaim> desiredPvcs = zkCluster.generatePersistentVolumeClaims();
                     Future<List<PersistentVolumeClaim>> existingPvcsFuture = pvcOperations.listAsync(namespace, Labels.fromMap(zkCluster.getSelectorLabels()));
 
-                    maybeCleanPodAndPvc(res.result(), desiredPvcs, existingPvcsFuture).setHandler(resultFuture.completer());
+                    maybeCleanPodAndPvc(zkSetOperations, res.result(), desiredPvcs, existingPvcsFuture).setHandler(resultFuture.completer());
                 } else {
                     resultFuture.fail(res.cause());
                 }
@@ -1749,7 +1748,7 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
 
                     Future<List<PersistentVolumeClaim>> existingPvcsFuture = pvcOperations.listAsync(namespace, Labels.fromMap(kafkaCluster.getSelectorLabels()));
 
-                    maybeCleanPodAndPvc(sts, desiredPvcs, existingPvcsFuture).setHandler(resultFuture.completer());
+                    maybeCleanPodAndPvc(kafkaSetOperations, sts, desiredPvcs, existingPvcsFuture).setHandler(resultFuture.completer());
                 } else {
                     resultFuture.fail(res.cause());
                 }
@@ -1768,12 +1767,13 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
          * to not happen in parallel or have a risk that there will be several pods deleted at the same time (which can
          * affect availability). If multiple pods are marked for deletion, the next one will be deleted in the next loop.
          *
+         * @param stsOperator           StatefulSet Operator for managing stateful sets
          * @param sts                   StatefulSet which owns the pods which should be checked for deletion
          * @param desiredPvcs           The list of PVCs which should exist
          * @param existingPvcsFuture    Future which will return a list of PVCs which actually exist
          * @return
          */
-        Future<Void> maybeCleanPodAndPvc(StatefulSet sts, List<PersistentVolumeClaim> desiredPvcs, Future<List<PersistentVolumeClaim>> existingPvcsFuture)  {
+        Future<Void> maybeCleanPodAndPvc(StatefulSetOperator stsOperator, StatefulSet sts, List<PersistentVolumeClaim> desiredPvcs, Future<List<PersistentVolumeClaim>> existingPvcsFuture)  {
             if (sts != null) {
                 log.debug("{}: Considering manual cleaning of Pods for StatefulSet {}", reconciliation, sts.getMetadata().getName());
 
@@ -1806,7 +1806,7 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                                                 .filter(pvc -> pvc.getMetadata().getName().endsWith(podName))
                                                 .collect(Collectors.toList());
 
-                                        return cleanPodAndPvc(podName, deletePvcs, createPvcs);
+                                        return cleanPodAndPvc(stsOperator, sts, podName, deletePvcs, createPvcs);
                                     });
                         }
                     }
@@ -1817,49 +1817,76 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
         }
 
         /**
-         * This is an internal method which actually executes the deletion of the Pod and PVC. There is a lot of
-         * unpredictable behaviour here:
-         *     - The PVCs (sometimes) cannot be deleted before the pods is deleted.
-         *     - Sometimes the Pod is recreated before the PVC is actually deleted and doesn't create the new PVCs
+         * This is an internal method which actually executes the deletion of the Pod and PVC. This is a non-trivial
+         * since the PVC and the Pod are tightly coupled and one cannot be deleted without the other. Also, the
+         * StatefulSet controller will recreate the deleted Pod and make it hard to recreate the PVCs manually.
          *
          * To address these, we:
-         *     1. Delete the PVC (= mark the PVC for deletion)
+         *     1. Delete the STS in non-cascading delete
          *     2. Delete the Pod
-         *     3. Reconcile the PVCs (either we create them or we just reconcile what the StatefulSet / Pods created)
-         *     4. Wait for Pod readiness
+         *     3. Delete the PVC
+         *     4. Wait for the Pod to be actually deleted
+         *     5. Wait for the PVCs to be actually deleted
+         *     6. Recreate the PVCs
+         *     7. Recreate the STS (which will in turn recreate the Pod)
+         *     5. Wait for Pod readiness
          *
-         * @param podName Name of the pod which should be deleted
-         * @param deletePvcs The list of PVCs which should be deleted
-         * @param createPvcs The list of PVCs which should be recreated
+         * @param stsOperator       StatefulSet Operator for managing stateful sets
+         * @param sts               The current StatefulSet to which the cleaned pod belongs
+         * @param podName           Name of the pod which should be deleted
+         * @param deletePvcs        The list of PVCs which should be deleted
+         * @param createPvcs        The list of PVCs which should be recreated
          * @return
          */
-        Future<Void> cleanPodAndPvc(String podName, List<PersistentVolumeClaim> deletePvcs, List<PersistentVolumeClaim> createPvcs) {
+        Future<Void> cleanPodAndPvc(StatefulSetOperator stsOperator, StatefulSet sts, String podName, List<PersistentVolumeClaim> deletePvcs, List<PersistentVolumeClaim> createPvcs) {
             long pollingIntervalMs = 1_000;
             long timeoutMs = operationTimeoutMs;
 
-            List<Future> deleteResults = new ArrayList<>(deletePvcs.size());
-
-            for (PersistentVolumeClaim pvc : deletePvcs)    {
-                String pvcName = pvc.getMetadata().getName();
-                log.debug("{}: Deleting PVC {} for Pod {} based on {} annotation", reconciliation, pvcName, podName, AbstractScalableResourceOperator.ANNO_STRIMZI_IO_DELETE_POD_AND_PVC);
-                deleteResults.add(pvcOperations.reconcile(namespace, pvcName, null));
-            }
-
-            Future fut = CompositeFuture
-                    .join(deleteResults)
-                    .compose(ignored -> podOperations.reconcile(namespace, podName, null))
+            // We start by deleting the StatefulSet so that it doesn't interfere with the pod deletion process
+            // The deletion has to be non-cascading so that the other pods are not affected
+            Future<Void> fut = stsOperator.deleteAsync(namespace, sts.getMetadata().getName(), false)
                     .compose(ignored -> {
+                        // After the StatefulSet is deleted, we can delete the pod which was marked for deletion
+                        return podOperations.reconcile(namespace, podName, null);
+                    })
+                    .compose(ignored -> {
+                        // With the pod deleting, we can delete all the PVCs belonging to this pod
+                        List<Future> deleteResults = new ArrayList<>(deletePvcs.size());
+
+                        for (PersistentVolumeClaim pvc : deletePvcs)    {
+                            String pvcName = pvc.getMetadata().getName();
+                            log.debug("{}: Deleting PVC {} for Pod {} based on {} annotation", reconciliation, pvcName, podName, AbstractScalableResourceOperator.ANNO_STRIMZI_IO_DELETE_POD_AND_PVC);
+                            deleteResults.add(pvcOperations.reconcile(namespace, pvcName, null));
+                        }
+                        return CompositeFuture.join(deleteResults);
+                    })
+                    .compose(ignored -> {
+                        // The pod deletion just triggers it asynchronously
+                        // We have to wait for the pod to be actually deleted
+                        log.debug("{}: Checking if Pod {} has been deleted", reconciliation, podName);
+
+                        Future<Void> waitForDeletion = podOperations.waitFor(namespace, podName, pollingIntervalMs, timeoutMs, (ignore1, ignore2) -> {
+                            Pod deletion = podOperations.get(namespace, podName);
+                            log.trace("Checking if Pod {} in namespace {} has been deleted or recreated", podName, namespace);
+                            return deletion == null;
+                        });
+
+                        return waitForDeletion;
+                    })
+                    .compose(ignored -> {
+                        // Once the pod is deleted, the PVCs should delete as well
+                        // Faked PVCs on Minishift etc. might delete while the pod is running, real PVCs will not
                         List<Future> waitForDeletionResults = new ArrayList<>(deletePvcs.size());
 
                         for (PersistentVolumeClaim pvc : deletePvcs)    {
                             String pvcName = pvc.getMetadata().getName();
                             String uid = pvc.getMetadata().getUid();
 
-                            log.debug("{}: Checking if PVC {} for Pod {} has been deleted or recreated", reconciliation, pvcName, podName);
+                            log.debug("{}: Checking if PVC {} for Pod {} has been deleted", reconciliation, pvcName, podName);
 
                             Future<Void> waitForDeletion = pvcOperations.waitFor(namespace, pvcName, pollingIntervalMs, timeoutMs, (ignore1, ignore2) -> {
                                 PersistentVolumeClaim deletion = pvcOperations.get(namespace, pvcName);
-                                log.trace("Checking if {} {} in namespace {} has been deleted or recreated", pvc.getKind(), pvcName, namespace);
+                                log.trace("Checking if {} {} in namespace {} has been deleted", pvc.getKind(), pvcName, namespace);
                                 return deletion == null || (deletion.getMetadata() != null && !uid.equals(deletion.getMetadata().getUid()));
                             });
 
@@ -1869,6 +1896,8 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                         return CompositeFuture.join(waitForDeletionResults);
                     })
                     .compose(ignored -> {
+                        // Once everything was deleted, we can start recreating it.
+                        // First we recreate the PVCs
                         List<Future> createResults = new ArrayList<>(createPvcs.size());
 
                         for (PersistentVolumeClaim pvc : createPvcs)    {
@@ -1877,6 +1906,20 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                         }
 
                         return CompositeFuture.join(createResults);
+                    })
+                    .compose(ignored -> {
+                        // After the PVCs we recreate the StatefulSet which will in turn recreate the missing Pod
+                        // We cannot use the new StatefulSet here because there might have been some other changes for
+                        // which we are not yet ready - e.g. changes to off-cluster access etc. which might not work
+                        // without the CO creating some other infrastructure.
+                        // Therefore we use the old STS and just remove some things such as Status, ResourceVersion, UID
+                        // or self link. These will be recreated by Kubernetes after it is created.
+                        sts.getMetadata().setResourceVersion(null);
+                        sts.getMetadata().setSelfLink(null);
+                        sts.getMetadata().setUid(null);
+                        sts.setStatus(null);
+
+                        return stsOperator.reconcile(namespace, sts.getMetadata().getName(), sts);
                     })
                     .compose(ignored -> podOperations.readiness(namespace, podName, pollingIntervalMs, timeoutMs));