+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add ability to configure Image Pull Secrets for all pods in Cluster Operator
 * Support for SASL PLAIN mechanism in Kafka Connect and Mirror Maker (for use with non-Strimzi Kafka cluster)
 * Add resizing of persistent volumes
+* Allow to specify different storage class for every broker
 
 ## 0.11.0